Display format into Session instead of main
