+++ src/session.rs
@@ -1,6 +1,7 @@
 use crate::{
     event::Event,
     records::{EventRecord, SessionRecord},
+    Cli,
 };
 use chrono::{DateTime, Duration, FixedOffset};
 use std::{collections::VecDeque, net::IpAddr};
@@ -120,5 +121,79 @@ impl Session {
     /// Given by summing the total durations of all root traces.
     pub fn total_duration(&self) -> i64 {
         self.root_events.iter().map(|e| e.durations().0).sum()
+    }
+
+    pub fn display(&self, cli: Cli) {
+        // Print out the session info
+        println!("Session ID: {}", &self.id);
+        println!("{}", &self.started_at.to_rfc3339());
+        println!(
+            "{:15} ({}) -> {:15}",
+            &self.client,
+            &self.username.clone().unwrap_or_else(|| String::from("N/A")),
+            &self.coordinator
+        );
+        println!(
+            "Request Size:  {}",
+            &self
+                .request_size
+                .map(|rs| rs.to_string())
+                .unwrap_or_else(|| String::from("N/A"))
+        );
+        println!(
+            "Response Size: {}",
+            &self
+                .response_size
+                .map(|rs| rs.to_string())
+                .unwrap_or_else(|| String::from("N/A"))
+        );
+        println!("{}", &self.request);
+        println!("{:?}", &self.parameters);
+
+        // Calculations for the waterfall boxes
+        let s_end = self.total_duration();
+        let mut offset = 0i64;
+
+        let events = self.events();
+        let a_max_width = events
+            .iter()
+            .map(|(e, _)| e.activity_length())
+            .max()
+            .unwrap_or(0);
+        let max_depth = events.iter().map(|(_, depth)| *depth).max().unwrap_or(1);
+        let i_max_width = self.event_count().to_string().len();
+
+        // Headers
+        println!();
+        println!(
+            "{:i_max_width$} {:w_width$} {}",
+            "",
+            "waterfall chart",
+            crate::event_display_str(
+                &cli,
+                a_max_width,
+                "dur",
+                "node",
+                &format!("{:tree_width$}", "", tree_width = max_depth + 2),
+                "activity",
+                "event id",
+                "span id",
+                "parent span id",
+                "thread name",
+            ),
+            w_width = cli.waterfall_width + 2
+        );
+
+        for (i, (e, depth)) in events.iter().enumerate() {
+            println!(
+                "{:i_max_width$} {} {}",
+                i + 1,
+                e.waterfall(&cli, offset, s_end),
+                e.display(&cli, a_max_width, *depth, max_depth)
+            );
+
+            // Move the offset up for the next event
+            offset += e.durations().1;
+        }
     }
 }
